Migrate User route to User module
